@@ -8,14 +8,14 @@
     <name>Solace Event Management Agent - Solace Plugin</name>
     <description>Solace Event Management Agent - Solace Plugin</description>
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <solace-messaging-client.version>1.0.0</solace-messaging-client.version>
         <solclientj.version>10.0.0</solclientj.version>
-        <spring-boot-starter-webflux.version>2.7.6</spring-boot-starter-webflux.version>
-        <spring-boot-starter-test.version>2.7.6</spring-boot-starter-test.version>
+        <spring-boot-starter-webflux.version>3.1.2</spring-boot-starter-webflux.version>
+        <spring-boot-starter-test.version>3.1.2</spring-boot-starter-test.version>
         <jupiter.version>5.3.1</jupiter.version>
-        <camel.version>3.16.0</camel.version>
+        <camel.version>4.0.0-RC2</camel.version>
         <commons-collections4.version>4.4</commons-collections4.version>
         <snakeyaml.version>2.0</snakeyaml.version>
         <maven-shade-plugin.version>3.3.0</maven-shade-plugin.version>
@@ -41,6 +41,10 @@
                 <exclusion>
                     <groupId>org.yaml</groupId>
                     <artifactId>snakeyaml</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>ch.qos.logback</groupId>
+                    <artifactId>logback-classic</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
