@@ -1,8 +1,6 @@
 package com.solace.maas.ep.event.management.agent.config;
 
-import com.solace.maas.ep.event.management.agent.logging.CustomUnitOfWork;
 import org.apache.camel.CamelContext;
-import org.apache.camel.ExtendedCamelContext;
 import org.apache.camel.spring.boot.CamelContextConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
@@ -16,7 +14,7 @@ public class CamelContextConfig {
             @Override
             public void beforeApplicationStart(CamelContext context) {
                 context.setUseMDCLogging(true);
-                context.adapt(ExtendedCamelContext.class).setUnitOfWorkFactory(CustomUnitOfWork::new);
+                context.setStreamCaching(false);
             }
 
             @Override
